integ - setup envs, check, run tests, cleanup
